changed to float precision on collision handling

+++ GameEngine/src/Collision/CollisionHandler.h
@@ -6,16 +6,16 @@ class CollisionHandler
 {
     
 public:
-    static int DistanceSquared(float2 PositionA, float2 PositionB)
+    static float DistanceSquared(float2 PositionA, float2 PositionB)
     {
-        int deltaX = PositionB.X - PositionA.X;
-        int deltaY = PositionB.Y - PositionA.Y;
+        float deltaX = PositionB.X - PositionA.X;
+        float deltaY = PositionB.Y - PositionA.Y;
         return (deltaX * deltaX) + (deltaY * deltaY);
     }
     
     static bool Collided(float2 PositionA, int RadiusA, float2 PositionB, int RadiusB)
     {
-        const int DistaneSquared = DistanceSquared(PositionA, PositionB);
+        const float DistaneSquared = DistanceSquared(PositionA, PositionB);
         const int RadiiSum = RadiusA + RadiusB;
         
         return DistaneSquared <= (RadiiSum * RadiiSum);
